pipecat-examples/websocket/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Pipecat example - server"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "loguru",
    "python-dotenv",
    "uvicorn",
    "pipecat-ai[silero,websocket,google]>=0.0.102",
]

[dependency-groups]
dev = [
    "pyright>=1.1.404,<2",
    "ruff>=0.12.11,<1",
]

[tool.ruff]
line-length = 100
[tool.ruff.lint]
select = ["I"]
